test(storage): better errors on cleanup (#3611)

+++ src/storage/examples/src/lib.rs
@@ -661,7 +661,7 @@ async fn create_bucket_kms_key(
 }
 
 pub async fn cleanup_bucket(client: StorageControl, name: String) -> anyhow::Result<()> {
-    use google_cloud_gax::paginator::ItemPaginator;
+    use google_cloud_gax::{Result as GaxResult, paginator::ItemPaginator};
     use google_cloud_wkt::FieldMask;
 
     // Configure the bucket to be garbage collected. Some buckets are created by
@@ -673,13 +673,13 @@ pub async fn cleanup_bucket(client: StorageControl, name: String) -> anyhow::Res
         .get("integration-tests")
         .is_none_or(|v| v != "true")
     {
-        let mut current = current;
-        current
+        let mut updated = current.clone();
+        updated
             .labels
             .insert("integration-tests".to_string(), "true".to_string());
         if let Err(e) = client
             .update_bucket()
-            .set_bucket(current)
+            .set_bucket(updated)
             .set_update_mask(FieldMask::default().set_paths(["labels"]))
             .send()
             .await
@@ -709,27 +709,47 @@ pub async fn cleanup_bucket(client: StorageControl, name: String) -> anyhow::Res
                 .send(),
         );
     }
-    let _ = futures::future::join_all(pending).await;
+    if let Err(e) = futures::future::join_all(pending)
+        .await
+        .into_iter()
+        .collect::<GaxResult<Vec<_>>>()
+    {
+        tracing::error!("Error cleaning up objects in bucket {name}: {e:?}");
+    };
 
-    let mut pending = Vec::new();
-    let mut folders = client.list_managed_folders().set_parent(&name).by_item();
-    while let Some(item) = folders.next().await {
-        let Ok(folder) = item else {
-            continue;
-        };
-        pending.push(client.delete_managed_folder().set_name(folder.name).send());
-    }
-    let _ = futures::future::join_all(pending).await;
+    if current.hierarchical_namespace.is_some_and(|h| h.enabled) {
+        let mut pending = Vec::new();
+        let mut folders = client.list_managed_folders().set_parent(&name).by_item();
+        while let Some(item) = folders.next().await {
+            let Ok(folder) = item else {
+                continue;
+            };
+            pending.push(client.delete_managed_folder().set_name(folder.name).send());
+        }
+        if let Err(e) = futures::future::join_all(pending)
+            .await
+            .into_iter()
+            .collect::<GaxResult<Vec<_>>>()
+        {
+            tracing::error!("Error cleaning up managed folders in bucket {name}: {e:?}");
+        }
 
-    let mut pending = Vec::new();
-    let mut folders = client.list_folders().set_parent(&name).by_item();
-    while let Some(item) = folders.next().await {
-        let Ok(folder) = item else {
-            continue;
+        let mut pending = Vec::new();
+        let mut folders = client.list_folders().set_parent(&name).by_item();
+        while let Some(item) = folders.next().await {
+            let Ok(folder) = item else {
+                continue;
+            };
+            pending.push(client.delete_folder().set_name(folder.name).send());
+        }
+        if let Err(e) = futures::future::join_all(pending)
+            .await
+            .into_iter()
+            .collect::<GaxResult<Vec<_>>>()
+        {
+            tracing::error!("Error cleaning up folders in bucket {name}: {e:?}");
         };
-        pending.push(client.delete_folder().set_name(folder.name).send());
     }
-    let _ = futures::future::join_all(pending).await;
 
     let mut pending = Vec::new();
     let mut caches = client.list_anywhere_caches().set_parent(&name).by_item();
@@ -739,7 +759,13 @@ pub async fn cleanup_bucket(client: StorageControl, name: String) -> anyhow::Res
         };
         pending.push(client.disable_anywhere_cache().set_name(cache.name).send());
     }
-    let _ = futures::future::join_all(pending).await;
+    if let Err(e) = futures::future::join_all(pending)
+        .await
+        .into_iter()
+        .collect::<GaxResult<Vec<_>>>()
+    {
+        tracing::error!("Error cleaning up caches in bucket {name}: {e:?}");
+    };
 
     client.delete_bucket().set_name(&name).send().await?;
     Ok(())
